+++ src/components/navigation/HomeNavbar.tsx
@@ -3,7 +3,7 @@ import Link from "next/link";
 import { type FC } from "react";
 
 import Searchbar from "../Searchbar";
-import Dropdown from "../Dropdown";
+import Dropdown from "../menus/Dropdown";
 
 const HomeNavbar: FC = () => {
   return (
